+++ src/main/java/model/card/Rank.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 import java.util.Arrays;
