+++ src/test/java/model/participant/BlackJackTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import java.util.Map;
 import model.BlackJack;
-import model.Card;
+import model.card.Card;
 import model.Participants;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
