+++ src/test/java/model/participant/DealerTest.java
@@ -3,7 +3,7 @@ package model.participant;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import model.Card;
+import model.card.Card;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
